requirements/local.txt
-r common.txt
black==23.7.0
ruff==0.0.287
mypy==1.5.1
pre-commit==3.3.3
pytest-django==4.5.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
